hemp-inventory-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.1"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
httpx = "^0.28.1"
pyjwt = "^2.11.0"
python-dotenv = "^1.2.2"
aiosqlite = "^0.22.1"
bcrypt = "^5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
apscheduler = "^3.11.2"
aiosmtplib = "^5.1.0"
openai = "^2.26.0"
pillow = "^12.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
